Make main.rb to be responsible for only starting the code
